fixing date

+++ application/business_logic.py
@@ -137,7 +137,7 @@ def generate_list_of_adults_not_completed_health_check(no_days):
         adults_waiting = response.record
         for adult in adults_waiting:
             email_sent = datetime.strptime(adult['email_resent_timestamp'][:10], '%Y-%m-%d')
-            no_days = int(no_days)
+            no_days = float(no_days)
             if now - relativedelta(days=no_days) <= email_sent and not adult[sent_field]:
                 adults_to_remind.append(adult)
     elif response.status_code == 404:
